requirements.txt
streamlit
opencv-python
pytesseract
pandas
openpyxl
numpy
Pillow
